Fixed some warnings

+++ src/Detector/tensorrt_yolo/yolov5.h
@@ -22,10 +22,10 @@ public:
 		b.y1 = y - bh / 2;
 		b.y2 = y + bh / 2;
 
-		b.x1 = clamp(b.x1, 0, netW);
-		b.x2 = clamp(b.x2, 0, netW);
-		b.y1 = clamp(b.y1, 0, netH);
-		b.y2 = clamp(b.y2, 0, netH);
+		b.x1 = clamp(b.x1, 0, static_cast<float>(netW));
+		b.x2 = clamp(b.x2, 0, static_cast<float>(netW));
+		b.y1 = clamp(b.y1, 0, static_cast<float>(netH));
+		b.y2 = clamp(b.y2, 0, static_cast<float>(netH));
 
 		return b;
 	}
